refactor(notes): mueve modules/notes a features/notes con layout ui/model/api/hooks
