+++ src/actions/archive/index.ts
@@ -1,4 +1,2 @@
 export * from './getArchive'
 export * from './saveArchive'
-export * from './getArchivedNotes'
-export * from './saveArchivedNotes'
