--- src/actions/notes/index.ts
@@ -1,2 +1,0 @@
-export * from './getNotes'
-export * from './saveNotes'
