+++ src/features/notes/api/libraryOfArchivedNotesRepository.ts
@@ -1,4 +1,4 @@
-import { LibraryOfArchivedNotes } from './libraryOfArchivedNotes'
+import { LibraryOfArchivedNotes } from '../model/libraryOfArchivedNotes'
 
 export interface LibraryOfArchiveNotesRepository {
 	save(notes: LibraryOfArchivedNotes, boardId: string): Promise<void>
