+++ src/features/notes/api/nextjsNotesRepository.ts
@@ -3,12 +3,12 @@ import { NotesRepository } from './notesRepository'
 
 export default class NextjsNotesRepository implements NotesRepository {
 	async save(notes: Notes, boardId: string): Promise<void> {
-		const { saveNotes } = await import('@/actions/notes')
+		const { saveNotes } = await import('./saveNotes')
 		await saveNotes({ boardId, notes })
 	}
 
 	async getAll(boardId: string): Promise<Notes> {
-		const { getNotes } = await import('@/actions/notes')
+		const { getNotes } = await import('./getNotes')
 		const notes = await getNotes({ boardId })
 		return notes ?? defaultNotes
 	}
