+++ src/features/notes/api/saveNotes.ts
@@ -1,7 +1,7 @@
 'use server'
 
 import { prisma } from '@/shared/lib/prisma'
-import type { Notes } from '@/modules/notes/model/notes'
+import type { Notes } from '../model/notes'
 import { requireBoardAccess } from '@/shared/lib/serverAuth'
 
 export async function saveNotes({
