+++ src/features/notes/hooks/useNotesQuery.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
-import { fetchNotes, saveNotes } from '../repository/notesRepositoryFactory'
+import { fetchNotes, saveNotes } from '../api/notesRepositoryFactory'
 import { useSession } from '@/auth/hooks/useSession'
 import { defaultNotes, Notes } from '../model/notes'
 import { useBoardId } from '@/auth/state/store'
