+++ src/features/notes/index.ts
@@ -1,0 +1,4 @@
+export { default as Notes } from './ui/Notes'
+export { NoteInput } from './ui/NoteInput'
+export { default as LibraryOfArchivedNotes } from './ui/LibraryOfArchivedNotes'
+export { defaultNotes } from './model/notes'
