+++ src/features/notes/ui/NoteInput.tsx
