+++ src/shared/ui/organisms/Header.tsx
@@ -24,7 +24,7 @@ import { LanguageToggle } from '../../../modules/LanguageToggle/LanguageToggle'
 import { useTranslation } from 'react-i18next'
 import LogInAndLogOutMenuItem from '../../../modules/LanguageToggle/LogInAndLogOutMenuItem'
 import { useSession } from '@/auth/hooks/useSession'
-import Notes from '@/modules/notes/Notes'
+import { Notes } from '@/features/notes'
 import { useTheme } from '@/shared/hooks/useTheme'
 import { useLastDurationPeriod } from '@/modules/UsageHistory/hooks/useLastDurationPeriod'
 import { useTypeOfView } from '@/modules/TypeOfView/useTypeOfView'
