+++ src/views/BoardPage.tsx
@@ -12,7 +12,7 @@ import { USER_IS_IN } from '@/shared/ui/organisms/userIsIn'
 import { ListView } from '@/modules/TaskBoard/components/ListView'
 import { TableView } from '@/modules/TaskBoard/components/TableView'
 import { useTypeOfView } from '@/modules/TypeOfView/useTypeOfView'
-import { NoteInput } from '@/modules/notes/components/NoteInput'
+import { NoteInput } from '@/features/notes'
 import { useTaskBoardQuery } from '@/modules/TaskBoard/hooks/useTaskBoardQuery'
 import { useReminder } from '@/modules/TaskBoard/components/Reminder/hooks/useReminder'
 import { useSession } from '@/auth/hooks/useSession'
